Move RecurringSchedule.ts to app/models.
